requirements.txt
Flask==2.0.1
Werkzeug==2.0.3 # Added to ensure compatibility with Flask 2.0.1
requests==2.26.0
python-dotenv==0.19.0
